Break events.js into multiple files
