+++ core/contextmenu.js
@@ -30,6 +30,7 @@
  */
 goog.provide('Blockly.ContextMenu');
 
+goog.require('Blockly.Events.BlockCreate');
 goog.require('Blockly.utils');
 goog.require('Blockly.utils.uiMenu');
 
@@ -38,6 +39,7 @@ goog.require('goog.events');
 goog.require('goog.style');
 goog.require('goog.ui.Menu');
 goog.require('goog.ui.MenuItem');
+goog.require('goog.userAgent');
 
 
 /**
