+++ core/gesture.js
@@ -30,7 +30,7 @@ goog.provide('Blockly.Gesture');
 goog.require('Blockly.BlockAnimations');
 goog.require('Blockly.BlockDragger');
 goog.require('Blockly.constants');
-goog.require('Blockly.Events');
+goog.require('Blockly.Events.Ui');
 goog.require('Blockly.FlyoutDragger');
 goog.require('Blockly.scratchBlocksUtils');
 goog.require('Blockly.Tooltip');
