+++ core/variable_map.js
@@ -26,7 +26,8 @@
 
 goog.provide('Blockly.VariableMap');
 
-goog.require('Blockly.VariableModel');
+goog.require('Blockly.Events.VarDelete');
+goog.require('Blockly.Events.VarRename');
 
 /**
  * Class for a variable map.  This contains a dictionary data structure with
