+++ core/warning.js
@@ -27,6 +27,7 @@
 goog.provide('Blockly.Warning');
 
 goog.require('Blockly.Bubble');
+goog.require('Blockly.Events.Ui');
 goog.require('Blockly.Icon');
